example cmake
