+++ src/yolov4-tiny/CMakeLists.txt
@@ -1,0 +1,14 @@
+cmake_minimum_required(VERSION 3.10)
+
+project(yolov4-tiny)
+
+find_package(PkgConfig)
+pkg_check_modules(OPENCV REQUIRED opencv4)
+include_directories(${OPENCV_INCLUDE_DIRS})
+
+include_directories(${CMAKE_SOURCE_DIR})
+
+include_directories(/opt/tflite-2.4.2/include)
+
+add_library(yolov4-tiny yolov4-tiny.cc)
+target_link_libraries(yolov4-tiny ${OPENCV_STATIC_LDFLAGS})
